Splitting ClockifyClient.cs to partial classes.
